% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conservation_problem.R
\name{ConservationProblem-class}
\alias{ConservationProblem-class}
\alias{ConservationProblem}
\title{Conservation problem class}
\description{
This class is used to represent a data instance for the multi-action
conservation planning problem. This includes several methods for obtaining
instance information. It is created using the \code{\link{problem}} function.
}
\section{Fields}{
 \describe{

  \item{$data}{\code{list} object containing data.}

  }
}

\section{Methods}{
 \describe{ \item{$getActionsAmount( )}{return
  \code{integer} number of possible actions.}

  \item{$getData(\code{character} name)}{return an \code{\link{data.frame}}
  object stored in the \code{data} field with the corresponding \code{name}.
  The argument \code{name} indicates the name of arguments of the
  \code{problem} function ("pu", "features", "rij", "threats", "sensitivity"
  or "boundary").}

  \item{$getFeatureAmount( )}{return \code{integer} number of features.}

  \item{$getFeatureNames( )}{return \code{character} names of features.}

  \item{$getPlanningUnitCosts( )}{return the \code{numeric}
  \code{\link{vector}} cost of allocating each planning unit.}

  \item{$getPlanningUnitsAmount( )}{return \code{integer} number of planning
  units.}

  \item{$getThreatCosts( )}{return the \code{numeric} \code{\link{vector}}
  cost of actions each planning unit and threat.}

  \item{$getThreatNames( )}{return \code{character} names of threats.}

  \item{$getThreatsAmount( )}{return \code{integer} number of threats.}

  \item{$print( )}{print basic information of the data instance.}

  \item{$show( )}{call print method.}

  }
}

\examples{
## Examples of how to use the methods of a ConservationProblem class object.

## Load data
data(example_pu_data, example_features_data, example_rij_data, example_threats_data, example_sensitivity_data, example_boundary_data)

## Create data instance
problem_data <- problem(
  pu = example_pu_data, features = example_features_data, dist_features = example_dist_features__data,
  threats = example_threats_data, dist_therats = example_dist_threats_data, sensitivity = example_sensibility_data,
  boundary = example_boundary_data
)

## Use class methods
problem_data$getData("features")

problem_data$getFeatureAmount()

problem_data$getFeatureNames()

problem_data$getPlanningUnitCosts()

problem_data$getPlanningUnitsAmount()

problem_data$getThreatCosts()

problem_data$getThreatNames()

problem_data$getThreatsAmount()

problem_data$print()
}
