% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getActions.R
\name{getActions}
\alias{getActions}
\title{Get actions}
\usage{
getActions(x, format = "large")
}
\arguments{
\item{x}{\linkS4class{solution} or \linkS4class{portfolio} object.}

\item{format}{\code{string}. Output format of the action matrix. \code{"wide"} format
shows one column per action, while \code{"large"} format shows three columns: pu,
threat and solution.}
}
\value{
\link{data.frame}.
}
\description{
Provides the selected actions in a solution.
}
\examples{
# set seed for reproducibility
set.seed(14)

## Load data
data(sim_pu_data, sim_features_data, sim_dist_features_data,
sim_threats_data, sim_dist_threats_data, sim_sensitivity_data,
sim_boundary_data)

## Create data instance
problem_data <- problem(
  pu = sim_pu_data, features = sim_features_data, dist_features = sim_dist_features_data,
  threats = sim_threats_data, dist_threats = sim_dist_threats_data, sensitivity = sim_sensitivity_data,
  boundary = sim_boundary_data
)

## Create optimization model
problem_model <- minimizeCosts(x = problem_data, blm = 1)

## Solve the optimization model
s <- solve(a = problem_model, solver = "gurobi", gap_limit = 0.01, output_file = FALSE)

# get actions information in large format
actions <- getActions(s, format = "large")
head(actions)

# get actions information in wide format
actions <- getActions(s, format = "wide")
head(actions)

}
