% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPlanningUnitsConnectivity.R
\name{getPlanningUnitsConnectivity}
\alias{getPlanningUnitsConnectivity}
\title{Get connectivity of planning units}
\usage{
getPlanningUnitsConnectivity(x)
}
\arguments{
\item{x}{\code{Solution-class} or \code{Portfolio-class} object.}
}
\value{
\link{numeric}.
}
\description{
Provides the connectivity value of all planning units in a solution.
}
\details{
The planning unit connectivity is calculated as the sum of all
boundaries unreached by the planning units in the solution. This can be expressed
mathematically for a set of planning units
\eqn{I} indexed by \eqn{i} and \eqn{j}, and
a set of threats \eqn{K} indexed by \eqn{k} as:

\deqn{
\sum_{i \in I}\sum_{j \in I} w_{i} (1 - w_{j})cv_{ij}
}

Here, \eqn{w_{i}} is the decisions variable that specify whether the planning
unit \eqn{i} has been selected (1) or not (0), \eqn{cv_{ij}}
is the large of boundary that share the planning units \eqn{i} and \eqn{j}.
If is a \code{recovery} model,
then all planning unit selected have at least one action to perform on it.
More information about of recovery models in
\code{minimizeCosts()} and \code{maximizeBenefits()} functions.
}
\examples{
# set seed for reproducibility
set.seed(14)

## Load data
data(sim_pu_data, sim_features_data, sim_dist_features_data,
sim_threats_data, sim_dist_threats_data, sim_sensitivity_data,
sim_boundary_data)

## Create data instance
problem_data <- problem(
  pu = sim_pu_data, features = sim_features_data, dist_features = sim_dist_features_data,
  threats = sim_threats_data, dist_threats = sim_dist_threats_data, sensitivity = sim_sensitivity_data,
  boundary = sim_boundary_data
)

## Create optimization model
problem_model <- minimizeCosts(x = problem_data, blm = 1)

## Solve the optimization model
s <- solve(a = problem_model, solver = "gurobi", gap_limit = 0.01, output_file = FALSE)

# get connectivity of units
getUnitConnectivity(s)

}
