% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStatus.R
\name{getStatus}
\alias{getStatus}
\title{Get status}
\usage{
getStatus(x)
}
\arguments{
\item{x}{\linkS4class{solution} or \linkS4class{portfolio} object.}
}
\value{
\link{character}.
}
\description{
Provides the status of solver at the end of the optimization period.
}
\details{
\code{GetStatus}can have five states:
\enumerate{
\item \emph{Optimal solution (according to gap tolerance)} : When the resolution of the model stop
when the quality of the solution (\emph{gap}) is less than or equal to
\code{gap_limit} (parameter of the \code{solve} function).
\item \emph{No solution (model was proven to be infeasible or unbounded)}: When the model is infeasible.
\item \emph{Feasible solution (according to time limit)}: When the resolution of the model
stops when a \code{time_limit} has been reached finding a feasible solution
(parameter of the \code{solve} function).
\item \emph{No solution (according to time limit)}: When the resolution of the model
stops when a \code{time_limit} has been reached without finding a feasible solution
(parameter of the \code{solve} function).
\item \emph{First feasible solution}: When the resolution of the model stops when it has
found the first feasible solution (\code{solution_limit = TRUE} parameter in \code{solve} function).
\item \emph{No solution information is available}: For any other case.
}
}
\examples{
# set seed for reproducibility
set.seed(14)

## Load data
data(sim_pu_data, sim_features_data, sim_dist_features_data,
sim_threats_data, sim_dist_threats_data, sim_sensitivity_data,
sim_boundary_data)

## Create data instance
problem_data <- problem(
  pu = sim_pu_data, features = sim_features_data, dist_features = sim_dist_features_data,
  threats = sim_threats_data, dist_threats = sim_dist_threats_data,
  sensitivity = sim_sensitivity_data, boundary = sim_boundary_data
)

## Create optimization model
problem_model <- minimizeCosts(x = problem_data, blm = 1)

## Solve the optimization model
s <- solve(a = problem_model, solver = "gurobi", gap_limit = 0.01, output_file = FALSE)

## get status of solution
getStatus(s)

}
