% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTotalConnectivity.R
\name{getTotalConnectivity}
\alias{getTotalConnectivity}
\title{Get total connectivity}
\usage{
getTotalConnectivity(x)
}
\arguments{
\item{x}{\linkS4class{solution} or \linkS4class{portfolio} object.}
}
\value{
\link{numeric}.
}
\description{
Provides the sum of connectivity (between actions and between planning units)
in a planning exercise. This is the sum of value returned from \code{getActionsConnectivity()} function and
the value returned from \code{getPlanningUnitsConnectivity()} function.
}
\examples{
# set seed for reproducibility
set.seed(14)

## Load data
data(sim_pu_data, sim_features_data, sim_dist_features_data,
sim_threats_data, sim_dist_threats_data, sim_sensitivity_data,
sim_boundary_data)

## Create data instance
problem_data <- problem(
  pu = sim_pu_data, features = sim_features_data, dist_features = sim_dist_features_data,
  threats = sim_threats_data, dist_threats = sim_dist_threats_data,
  sensitivity = sim_sensitivity_data, boundary = sim_boundary_data
)

## Create optimization model
problem_model <- minimizeCosts(x = problem_data, blm = 1)

## Solve the optimization model
s <- solve(a = problem_model, solver = "gurobi", gap_limit = 0.01, output_file = FALSE)

# get total connectivity of a solution
getTotalConnectivity(s)

}
\seealso{
\code{getActionsConnectivity()} and \code{getPlanningUnitsConnectivity()}.
}
