% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maximizeBenefits.R
\name{maximizeBenefits}
\alias{maximizeBenefits}
\title{Create mathematical model: maximize benefits}
\usage{
maximizeBenefits(
  x,
  blm = 0,
  budget = 0,
  curve = 3,
  segments = 3,
  recovery = TRUE
)
}
\arguments{
\item{x}{\linkS4class{conservationProblem} object. Data used in a problem of
prioritization of multiple conservation actions. This object must be created using the
\code{\link[=problem]{problem()}} function.}

\item{blm}{\code{numeric}. Penalty factor associated to the spatial fragmentation of planning
units, similar to Boundary Length Modifier (BLM) in \emph{Marxan}. This argument
only has an effect when the \code{boundary} is available.}

\item{budget}{\code{numeric}. Maximum budget to perform actions.}

\item{curve}{\code{integer}. Type of continuous curve used to represent benefit expression. It can
be a linear (\code{1}), quadratic (\code{2}) or cubic (\code{3}) function.
See \strong{Details} for more information.}

\item{segments}{\code{integer}. Number of segments (\code{1}, \code{2}, or \code{3}) used to approximate the non-linear
expression (\code{curve}) in the calculate benefits. See \strong{Details} for more information.}

\item{recovery}{\code{logical}. Indicates if it is a recovery (\code{TRUE}) or conservation problem (\code{FALSE}).
A recovery problem assumes no benefits for the features by incorporating units to the solution
where threats are not found (i.e. no co-occurrence between the feature and
its threats on that site).}
}
\value{
An object of class \linkS4class{optimizationProblem}.
}
\description{
Create an optimization model that maximize the sum of benefits of
all features without exceeding a budget limit for the multi-action conservation
planning problem. Following the mathematical formulations used in
Salgado-Rojas \emph{et al.} (2020). More information about mathematical formulation
in detail section.
}
\details{
The maximize benefits model seeks to find the set of actions that
maximizes the sum of benefits of all features, while the cost of performing
actions does not exceed a certain budget.

This model can be expressed mathematically for a set of planning units
\eqn{I} indexed by \eqn{i} a set of features \eqn{S} indexed by \eqn{s}, and
a set of threats \eqn{K} indexed by \eqn{k} as:

\deqn{
\max \space \sum_{i \in I}\sum_{s \in S_i} b_{is} r_{is} \\
\mathit{s.t.} \\
\sum_{i \in I} \sum_{k \in K_i} x_{ik} c_{ik} \leq budget
}

Here, \eqn{x_{ik}} is the decisions variable that specify whether action to abate the
threat \eqn{k} in the planning unit \eqn{i} has been selected (1) or not (0), \eqn{c_{ik}}
is the cost of do the action to abate the threat \eqn{k} in the planning unit \eqn{i},
\eqn{b_{is}} is the benefit of the feature \eqn{s} in the planning unit \eqn{i}
after doing actions on it (value between 0 and 1), \eqn{r_{is}} is the amount of
feature \eqn{s} in planning unit \eqn{i}.
And, the \eqn{t_s} is the target for feature \eqn{s}. The first term is the objective
function and the second is the set of constraints.

In the case of working with the presence/absence of threats (binary intensities),
we defined the benefit as a measure of the number of actions taken against the threats that
affect said feature with respect to all possible actions to do. However, this may
underestimate the true value of benefits. For this, we use a power (\code{curve} parameter)
to raise the expression and thus seek to perform a higher density of actions per site.
Because we work with linear models, we use the piecewise linearization strategy to
work with this expression. The \code{segments} parameter indicates how well the expression
approximates. A higher number implies a better approximation but increases the resolution complexity.
For more information on its calculation, see the
\href{https://prioriactions.github.io/prioriactions/reference/getBenefit.html}{getBenefit}
reference.

Connectivity parameters (\code{blm} and \code{blm_actions}) manipulate the
spatial fragmentation of planning units and/or action management to improve
the compactness of the reserve solutions.
}
\examples{
## This example uses input files included into package.

## set seed for reproducibility
set.seed(14)

## Load data
data(sim_pu_data, sim_features_data, sim_dist_features_data,
sim_threats_data, sim_dist_threats_data, sim_sensitivity_data,
sim_boundary_data)

## Create data instance
problem_data <- problem(
  pu = sim_pu_data, features = sim_features_data, dist_features = sim_dist_features_data,
  threats = sim_threats_data, dist_threats = sim_dist_threats_data, sensitivity = sim_sensitivity_data,
  boundary = sim_boundary_data
)

## Create optimization model
problem_model <- maximizeBenefits(x = problem_data, blm = 1, budget = 100)
}
\seealso{
For more information regarding the arguments
\code{curve} and \code{segments}, see the supplementary material
of Salgado-Rojas \emph{et al.} (2020), which can
be found online at \url{https://doi.org/10.1016/j.ecolmodel.2019.108901}.
}
