% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print}
\alias{print}
\alias{print.ConservationProblem}
\alias{print.OptimizationProblem}
\alias{print.Solution}
\title{Print}
\usage{
\method{print}{ConservationProblem}(x, ...)

\method{print}{OptimizationProblem}(x, ...)

\method{print}{Solution}(x, ...)
}
\arguments{
\item{x}{Any object.}

\item{...}{Not used.}
}
\value{
None.
}
\description{
Display information about an object.
}
\examples{
a <- 1:4
print(a)
}
\seealso{
\code{\link[base]{print}}.
}
