% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{show}
\alias{show}
\alias{show,}
\alias{ConservationProblem-method}
\alias{OptimizationProblem-method}
\alias{Solution-method}
\alias{show.ConservationProblem}
\alias{show.OptimizationProblem}
\alias{show.Solution}
\alias{show.Portafolio}
\title{Show}
\usage{
\method{show}{ConservationProblem}(x, ...)

\method{show}{OptimizationProblem}(x, ...)

\method{show}{Solution}(x, ...)

\method{show}{Portafolio}(x, ...)
}
\arguments{
\item{x}{Any object.}

\item{...}{Not used.}
}
\value{
None.
}
\description{
Display information about an object.
}
\examples{
a <- 1:4
show(a)
}
\seealso{
\code{\link[methods]{show}}.
}
