% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalBudget.R
\name{evalBudget}
\alias{evalBudget}
\title{Evaluate multiple budget values}
\usage{
evalBudget(data = list(), ...)
}
\arguments{
\item{data}{\code{list}. Input data list for \code{problem()} function.}

\item{...}{arguments inherited from \code{problem()}, \code{maximizeBenefits()},
and \code{solve()} functions.}

\item{budget}{\code{numeric}. Values of budget to verify. more than one value is
needed.}
}
\value{
An object of class \linkS4class{portfolio}.
}
\description{
Provides multiple solutions for different values of budgets. This function
assumes that you are working with the \emph{maximizeBenefits} model. Like
\code{prioriactions()} function, It inherits all arguments from \code{problem()},
\code{maximizeBenefits()} and \code{solve()}.
}
\details{
\code{evalBudget()} creates and solves multiple multi-actions planning
problems for different values of budgets. You can do this by manually running
\code{prioriactions()} function with these different budget values (i.e., running
once by budget). However, the \code{evalBudget()} function has two advantages
over their counterpart: 1) it is more efficient to create the models.
This is because the model is once created and then updated with the
new information; 2) the output is a portfolio object, which allows
obtaining information about the group of solutions, including, all \emph{get}
functions and also different types of plots.
}
\examples{
# set seed for reproducibility
set.seed(14)

## Load data
inputs <- list(sim_pu_data, sim_features_data, sim_dist_features_data,
sim_threats_data, sim_dist_threats_data, sim_sensitivity_data,
sim_boundary_data)

## Create model and solve
port <- evalBudget(data = inputs, budget = c(1, 10, 50, 100), time_limit = 50, output_file = FALSE)

plot(port)

}
