% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalTarget.R
\name{evalTarget}
\alias{evalTarget}
\title{sensitivityAnalysisBlm}
\usage{
evalTarget(data = list(), prop = 1, ...)
}
\arguments{
\item{x}{Object of class \code{\link[=ConservationProblem-class]{ConservationProblem-class()}} that
specifies the basic data used in a problem of prioritization of multiple
conservation actions. This object must be created using the
\code{\link[=problem]{problem()}} function.}

\item{blm}{A \code{numeric} value that indicates the penalty factor
associated to the spatial fragmentation of planning units, similar to
\strong{Boundary Length Modifier (BLM)} in \emph{Marxan}. This argument
only has an effect when the \code{\link[=bound]{bound()}} argument of
\code{\link[=problem]{problem()}} function is a \code{data.frame} object. \strong{The
default argument is zero}.}

\item{blm_actions}{A \code{numeric} value that indicates the penalty factor
associated to the spatial fragmentation of actions. \strong{The default
argument is zero}.}

\item{curve}{An \code{integer} value that selects the type of continuous
curve that will represent the expression (linear or non-linear) associated
with a specific constraint in this model. Therefore, the curve can
represent a linear (1), quadratic (2) or cubic (3) function. \strong{The
default argument is 3 and it is not recommended to change this value unless
you have advanced knowledge of the linearization of mathematical model}.}

\item{segments}{An \code{integer} value that selects the number of
segments (1, 2 or 3) that will have the \emph{piecewise linear function}
in charge of approximating the non-linear expression contained in a
specific constraint of this model. \strong{The default argument is 3 and it
is not recommended to change this value unless you have advanced knowledge
of the linearization of mathematical model}.}
}
\value{
An object of class \code{\link[=OptimizationProblem-class]{OptimizationProblem-class()}}.
}
\description{
Create an optimization model for the multi-action conservation
planning problem, following the mathematical formulations used in
Salgado-Rojas \emph{et al.} (2020). This function is used to specify model
configuration parameters related to connectivity issues and its internal
\emph{modus operandi}.

Connectivity parameters (\code{blm} and \code{blm_actions}) manipulate the
spatial fragmentation of planning units and/or action management to improve
the compactness of the reserve solutions. Likewise, the other parameters
(\code{curve} and \code{segments}) affect the linearization strategy
used by the model internally in order to be solved under a linear
programming approach. \strong{It is not recommended to modify the default
values of the later ones}.
}
\details{
\strong{Put details here! The details may include the mathematical
formulation of the optimization model associated with this conservation
problem and/or a rough description of the mathematical model, and/or what
happens when the parameters are set in one way or another.}
}
\examples{
## Create an optimization model for the multi-action conservation
## planning problem using a data instance that has been created in R.
## This example uses input files included into package.

## Load package
library(prioriactions)

## Load data
data(example_pu_data, example_features_data, example_dist_features_data, example_dist_threats_data, example_threats_data, example_sensitivity_data, example_bound_data)

## Create data instance
input <- list(
  pu = example_pu_data, features = example_features_data, dist_features = example_dist_features_data,
  threats = example_threats_data, dist_threats = example_dist_threats_data, sensitivity = example_sensitivity_data,
  bound = example_bound_data
)

## Create optimization model
portfolio <- evalTarget(data = input, prop = c(0.1, 0.2), output_file = FALSE, time_limit = 10)

}
\seealso{
For more information regarding the arguments \code{blm} and
\code{blm_actions}, see the \href{https://marxansolutions.org}{official \emph{Marxan} website} and the article by Salgado-Rojas \emph{et al.}
(2020), respectively. Also, for more information regarding the arguments
\code{curve} and \code{segments}, see the supplementary material
associated with the article by Salgado-Rojas \emph{et al.} (2020), which can
be found online at \url{https://doi.org/10.1016/j.ecolmodel.2019.108901}.
}
