% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPlanningUnits.R
\name{getPlanningUnits}
\alias{getPlanningUnits}
\title{Get planning units}
\usage{
getPlanningUnits(x)
}
\arguments{
\item{x}{\linkS4class{solution} or \linkS4class{portfolio} object.}
}
\value{
\link{data.frame}.
}
\description{
Provides the selected planning units in a solution. If is a \code{recovery} model,
then all planning unit selected have at least one action to perform on it.
More information about of recovery models in
\code{minimizeCosts()} and \code{maximizeBenefits()} functions.
}
\examples{
# set seed for reproducibility
set.seed(14)

## Load data
data(sim_pu_data, sim_features_data, sim_dist_features_data,
sim_threats_data, sim_dist_threats_data, sim_sensitivity_data,
sim_boundary_data)

## Create data instance
problem_data <- problem(
  pu = sim_pu_data, features = sim_features_data, dist_features = sim_dist_features_data,
  threats = sim_threats_data, dist_threats = sim_dist_threats_data,
  sensitivity = sim_sensitivity_data, boundary = sim_boundary_data
)

## Create optimization model
problem_model <- minimizeCosts(x = problem_data, blm = 1)

## Solve the optimization model
s <- solve(a = problem_model, solver = "gurobi", gap_limit = 0.01, output_file = FALSE)

# get planning units selected as solution
pu <- getPlanningUnits(s)
head(pu)

}
