% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\alias{plot.Portfolio}
\title{Plot}
\usage{
\method{plot}{Portfolio}(x, type = "connectivity_vs_cost", text = TRUE, ...)
}
\arguments{
\item{x}{Any object.}

\item{...}{Not used.}
}
\value{
None.
}
\description{
Display information about an object.
}
\seealso{
\code{\link[graphics:plot.default]{graphics::plot()}}.
}
