% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print}
\alias{print}
\alias{print.Data}
\alias{print.OptimizationProblem}
\alias{print.Solution}
\alias{print.Portfolio}
\title{Print}
\usage{
\method{print}{Data}(x, ...)

\method{print}{OptimizationProblem}(x, ...)

\method{print}{Solution}(x, ...)

\method{print}{Portfolio}(x, ...)
}
\arguments{
\item{x}{Any object.}

\item{...}{Not used.}
}
\value{
None.
}
\description{
Displays information about an object.
}
\seealso{
\code{\link[base:print]{base::print()}}.
}
