% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{prioriactions}
\alias{prioriactions}
\title{Create and solve multi-actions planning problems}
\usage{
prioriactions(data = list(), ...)
}
\arguments{
\item{data}{\code{list}. Input data list for \code{problem()} function.}

\item{...}{arguments inherited from \code{problem()}, \code{model()} and \code{solve()} functions.}
}
\value{
An object of class \linkS4class{solution}.
}
\description{
Create and solve a multi-actions planning problem. It can be used
instead of following the sequence of the \code{problem()}, \code{model()}
and \code{solve()} functions.
}
\examples{
## This example uses input files included into package.

## set seed for reproducibility
set.seed(14)

## Load data
data(sim_pu_data, sim_features_data, sim_dist_features_data,
sim_threats_data, sim_dist_threats_data, sim_sensitivity_data,
sim_boundary_data)

## Create data instance
input <- list(
  pu = sim_pu_data, features = sim_features_data, dist_features = sim_dist_features_data,
  threats = sim_threats_data, dist_threats = sim_dist_threats_data,
  sensitivity = sim_sensitivity_data, boundary = sim_boundary_data
)

## Create and solve optimization model
s <- prioriactions(data = input, name_model = "minimizeCosts",
                   blm = 0, output_file = FALSE, time_limit = 10)

print(s)

}
