% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problem.R
\name{problem}
\alias{problem}
\title{Multi-action conservation planning problem}
\usage{
problem(pu, features, rij, threats, sensibility, ...)

\S4method{problem}{data.frame}(pu, features, rij, threats, sensibility, bound = NULL)
}
\arguments{
\item{pu}{Object of class \code{\link{data.frame}} that specifies the planning units
to use in the reserve design exercise and their corresponding cost. It may be
desirable to exclude some planning units from the analysis, for example those outside
the study area. To exclude planning units you have two options:
(i) pre-process the \code{data.frame} before using it (i.e., \emph{manually} exclude planning units);
or (ii) set the \emph{status} of planning units to 3 not to include them. Each row corresponds
to a different planning unit, and each column corresponds to different information about the planning units.
The description of the columns is listed below.
\describe{
   \item{\code{"id"}}{\code{integer} unique identifier for each planning unit. This column is \strong{mandatory}.}
   \item{\code{"cost"}}{\code{numeric} cost of including each planning unit in the reserve system. This column is \strong{mandatory}.}
   \item{\code{"status"}}{\code{integer} indicating if each planning unit should not be locked in the solution (0) or if it should be \emph{locked-in} (2) or \emph{locked-out} (3) of the solution. This column is \strong{optional}.}
   }}

\item{features}{Object of class \code{\link{data.frame}} that specifies the conservation
features to consider in the reserve design exercise, such as their name and representation
requirements (i.e., \emph{targets}). Each row corresponds to a different feature and each column
corresponds to different information about the features. The description of the columns is listed below.
\describe{
   \item{\code{"id"}}{\code{integer} unique identifier for each conservation feature. This column is \strong{mandatory}.}
   \item{\code{"target"}}{\code{numeric} target amount for each conservation feature. This column is \strong{mandatory}.}
   \item{\code{"name"}}{\code{character} name for each conservation feature. This column is \strong{optional}.}
   }}

\item{rij}{Object of class \code{\link{data.frame}} that specifies the spatial distribution of conservation features across planning units.
\strong{It is recommended not to include cases where a feature does not occur in a planning unit (avoid rows with
\code{"amount"} equal to 0). Instead, it is recommended to omit these cases completely from the \code{\link{data.frame}}, for efficiency reasons.}
The description of the columns is listed below.
\describe{
   \item{\code{"pu"}}{\code{integer} \emph{id} of a planning unit where the conservation feature listed on the same row occurs. This column is \strong{mandatory}.}
   \item{\code{"species"}}{\code{integer} \emph{id} of each conservation feature. This column is \strong{mandatory}.}
   \item{\code{"amount"}}{\code{integer} indicating if the conservation feature occurring in the planning unit listed on the same row (1), or if it is not occurring (0). This column is \strong{mandatory}.}
   }}

\item{threats}{Object of class \code{\link{data.frame}} that specifies the spatial distribution of threats across planning units
and the corresponding cost of applying an action against these threats. \strong{It is recommended not to include cases where a threat does not occur
in a planning unit (avoid rows with \code{"amount"} equal to 0). Instead, it is recommended to omit these cases completely from the \code{\link{data.frame}}, for efficiency reasons.}
The description of the columns is listed below.
\describe{
   \item{\code{"pu"}}{\code{integer} \emph{id} of a planning unit where the threat listed on the same row occurs. This column is \strong{mandatory}.}
   \item{\code{"threats"}}{\code{integer} \emph{id} of each threat. This column is \strong{mandatory}.}
   \item{\code{"amount"}}{\code{integer} indicating if the threat occurring in the planning unit listed on the same row (1), or if it is not occurring (0). This column is \strong{mandatory}.}
   \item{\code{"cost"}}{\code{numeric} cost of applying an action to eliminate the threat in the planning unit listed in the same row. This column is \strong{mandatory}.}
   \item{\code{"status"}}{\code{integer} indicating if each action should not be locked in the solution (0) or if it should be \emph{locked-in} (2) or \emph{locked-out} (3) of the solution. This column is \strong{optional}.}
   }}

\item{sensibility}{Object of class \code{\link{data.frame}} that specifies the species-threats sensibility,
i.e., which threats affect the persistence of endangered conservation features (e.g., species).
\strong{It is recommended not to include cases where there is no species-threats sensibility
(avoid rows with \code{"amount"} equal to 0). Instead, it is recommended to omit these cases completely
from the \code{\link{data.frame}}, for efficiency reasons.}
The description of the columns is listed below.
   \describe{
   \item{\code{"species"}}{\code{integer} \emph{id} of each conservation feature. This column is \strong{mandatory}.}
   \item{\code{"threats"}}{\code{integer} \emph{id} of each threat. This column is \strong{mandatory}.}
   \item{\code{"amount"}}{\code{integer} indicating if the threat againsting the conservation feature listed on the same row (1), or if it is not againsting (0). This column is \strong{mandatory}.}
   }}

\item{bound}{Object of class \code{\link{NULL}} (default) or an object of class \code{\link{data.frame}}.
\code{NULL} indicates that planning units' boundaries (i.e., \emph{boundary data}) are not required for the multi-action conservation planning problem.
\code{data.frame} specifies the boundaries of the planning units, i.e., the spatial relationship between two units, such as the "length" of the shared boundary between them.
The description of the columns is listed below.
  \describe{
  \item{\code{"id1"}}{\code{integer} \emph{id} of each planning unit. This column is \strong{mandatory}.}
  \item{\code{"id2"}}{\code{integer} \emph{id} of each planning unit. This column is \strong{mandatory}.}
  \item{\code{"boundary"}}{\code{numeric} "length" of shared boundary between the planning units identified on the same row. This column is \strong{mandatory}.}
  }}
}
\value{
An object of class \code{\link{ConservationProblem-class}}.
}
\description{
Create a data instance for the multi-action conservation planning problem. This function is used to specify
the basic data used in a problem of prioritization of multiple conservation actions.
This includes, the spatial distribution of the planning units and their costs,
the features to be conserved (e.g. species, ecosystems) and their threats, and the
costs of implementing an action to remove such threats.
}
\details{
If you are familiar with using conservation planning software called
\emph{Marxan}, you should know that most of the arguments in this function
follow the conventions used by the Marxan input files. This means that,
\code{pu} corresponds to the data in the \strong{Planning Unit File} (typically called \code{"pu.dat"});
\code{features} corresponds to the data in the \strong{Conservation Feature File} (typically
called \code{"spec.dat"}); \code{rij} corresponds to the data in the \strong{Planning Unit versus
Conservation Feature File} (typically called \code{"puvspr.dat"}); and \code{bound} corresponds
to the data in the \strong{Boundary Length File} (typically called \code{"bound.dat"}).
}
\examples{
## Create a data instance for the multi-action conservation planning problem
## using data.frames that have been loaded into R. This example uses input files included
## into package.

## Load in planning unit data
pu_path <- system.file("extdata/input/pu.dat", package = "prioriactions")
pu_data <- data.table::fread(pu_path, data.table = FALSE)
head(pu_data)

## Load in feature data
spec_path <- system.file("extdata/input/spec.dat", package = "prioriactions")
spec_data <- data.table::fread(spec_path, data.table = FALSE)
head(spec_data)

## Load in planning unit vs feature data
puvspr_path <- system.file("extdata/input/puvspr.dat", package = "prioriactions")
puvspr_data <- data.table::fread(puvspr_path, data.table = FALSE)
head(puvspr_data)

## Load in the threats data
threats_path <- system.file("extdata/input/threats.dat", package = "prioriactions")
threats_data <- data.table::fread(threats_path, data.table = FALSE)
head(threats_data)

## Load in the sensibility data
sensibility_path <- system.file("extdata/input/sensibility.dat", package = "prioriactions")
sensibility_data <- data.table::fread(sensibility_path, data.table = FALSE)
head(sensibility_data)

## Load in the boundary data
bound_path <- system.file("extdata/input/bound.dat", package = "prioriactions")
bound_data <- data.table::fread(bound_path, data.table = FALSE)
head(bound_data)

## Create data instance
problem_data <- problem(
  pu = pu_data, features = spec_data, rij = puvspr_data,
  threats = threats_data, sensibility = sensibility_data,
  bound = bound_data
)

## Instance summary
print(problem_data)
}
\references{
\itemize{
\item Ball I, Possingham H, Watts, M. \emph{Marxan and relatives: software for spatial
conservation prioritization}. Spatial conservation prioritisation: quantitative
methods and computational tools 2009; 185<U+2012>195.
}
}
\seealso{
For more information on the correct format for \emph{Marxan} input data, see the
\href{https://marxansolutions.org}{official \emph{Marxan} website} and Ball \emph{et al.} (2009).
}
