% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{show}
\alias{show}
\alias{show,}
\alias{ConservationProblem-method}
\alias{OptimizationProblem-method}
\alias{Solution-method}
\alias{show.Data}
\alias{show.OptimizationProblem}
\alias{show.Solution}
\alias{show.Portfolio}
\title{Show}
\usage{
\method{show}{Data}(x, ...)

\method{show}{OptimizationProblem}(x, ...)

\method{show}{Solution}(x, ...)

\method{show}{Portfolio}(x, ...)
}
\arguments{
\item{x}{Any object.}

\item{...}{Not used.}
}
\value{
None.
}
\description{
Displays information about an object.
}
\examples{
a <- 1:4
show(a)
}
\seealso{
\code{\link[methods:show]{methods::show()}}.
}
