% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{show}
\alias{show}
\alias{show,Data-method}
\alias{show,OptimizationProblem-method}
\alias{show,Solution-method}
\alias{show,Portfolio-method}
\title{Show}
\usage{
\S4method{show}{Data}(x)

\S4method{show}{OptimizationProblem}(x)

\S4method{show}{Portfolio}(x)
}
\arguments{
\item{x}{Any object.}
}
\description{
Display information about an object.
}
\seealso{
\code{\link[methods:show]{methods::show()}}.
}
